+++ src/components/Form/FormValidation/FormValidation.tsx
@@ -1,15 +1,26 @@
 import { Box, Button, Slide } from "@tracktor/design-system";
+import { ReactNode } from "react";
+
+export interface RenderFormValidationParams {
+  disabled?: boolean;
+  isLoading?: boolean;
+}
 
 interface FormValidationProps {
   disabled?: boolean;
   isLoading?: boolean;
   isLastField?: boolean;
   readOnly?: boolean;
+  renderFormValidation?(params: RenderFormValidationParams): ReactNode;
 }
 
-const FormValidation = ({ disabled, isLoading, readOnly, isLastField }: FormValidationProps) => {
+const FormValidation = ({ disabled, isLoading, readOnly, isLastField, renderFormValidation }: FormValidationProps) => {
   if (readOnly || !isLastField) {
     return null;
+  }
+
+  if (renderFormValidation) {
+    return renderFormValidation({ disabled, isLoading });
   }
 
   return (