+++ src/features/TreegeConsumer/Standard/Standard.tsx
@@ -1,7 +1,7 @@
 import { Box, Stack } from "@tracktor/design-system";
-import { type CSSProperties, FormEvent } from "react";
+import { type CSSProperties, FormEvent, ReactNode } from "react";
 import FormSkeleton from "@/components/Feedback/FormSkeleton/FormSkeleton";
-import FormValidation from "@/components/Form/FormValidation";
+import FormValidation, { RenderFormValidationParams } from "@/components/Form/FormValidation";
 import FieldFactory from "@/components/Inputs/FieldFactory";
 import { TreegeConsumerProps } from "@/features/TreegeConsumer";
 import ChangeEventField from "@/types/ChangeEventField";
@@ -22,6 +22,7 @@ interface StandardProps {
   options?: TreegeConsumerProps["options"];
   handleChangeFormValue?(dataAttribute: ChangeEventField): void;
   handleSubmit?(event: FormEvent<HTMLFormElement>): void;
+  renderFormValidation?(params: RenderFormValidationParams): ReactNode;
 }
 
 const Standard = ({
@@ -37,6 +38,7 @@ const Standard = ({
   formCanBeSubmit,
   ignoreFields,
   options,
+  renderFormValidation,
 }: StandardProps) => (
   <Box onSubmit={handleSubmit} component="form" paddingX={15} paddingY={5} style={style}>
     <Stack spacing={4} direction="column" sx={{ "div:first-of-type hr": { display: "none" } }}>
@@ -58,7 +60,13 @@ const Standard = ({
         <FormSkeleton />
       )}
     </Stack>
-    <FormValidation disabled={!formCanBeSubmit} isLoading={isSubmitting} readOnly={readOnly} isLastField={isLastField} />
+    <FormValidation
+      disabled={!formCanBeSubmit}
+      isLoading={isSubmitting}
+      readOnly={readOnly}
+      isLastField={isLastField}
+      renderFormValidation={renderFormValidation}
+    />
   </Box>
 );